+++ src/account/operations/participation/voting.rs
@@ -2,16 +2,15 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use iota_client::{
+    api_types::plugins::participation::types::{
+        Participation, ParticipationEventId, Participations, PARTICIPATION_TAG,
+    },
     block::{
         output::{
             feature::{MetadataFeature, TagFeature},
             BasicOutputBuilder, Feature,
         },
         payload::TaggedDataPayload,
-    },
-    node_api::participation::types::{
-        participation::{Participation, Participations},
-        ParticipationEventId, PARTICIPATION_TAG,
     },
 };