+++ src/account/operations/transaction/high_level/burning_melting/destroy_alias.rs
@@ -6,7 +6,8 @@ use iota_client::{
     block::{
         address::{Address, AliasAddress},
         output::{
-            unlock_condition::AddressUnlockCondition, AliasId, BasicOutputBuilder, Output, OutputId, UnlockCondition,
+            unlock_condition::AddressUnlockCondition, AliasId, AliasTransition, BasicOutputBuilder, Output, OutputId,
+            UnlockCondition,
         },
     },
 };
@@ -41,7 +42,7 @@ impl AccountHandle {
                 &[Address::Alias(AliasAddress::new(alias_id))],
                 &output_data,
                 current_time,
-                true,
+                Some(AliasTransition::State),
             )? {
                 owned_outputs.push(output_data);
             }